Added some code to initialize a button/retvalue variables before they were accessed, to prevent an exception from being thrown.
Also changed the gamepad() main loop to use getGamepad() as opposed to
getGamepadNow(). This allows the tk event loop to pump, updating the
picture when a picture is taken. 
As far as I can tell, no benefits were gained from doing a busy wait
in this loop as opposed to a blocking wait.

I also changed the "done" variable to "doneSpeaking" to make the code
easier to understand.