Applied patch provide by Roberto Carrasco.
2) in ipre_s2_port.spin line 445 to 455
is best to use mutex or a temp var to enqueue the high and low bytes (see line 828)

3) in ipre_s2_port.spin line 709 (Get_Line_Ex):
missing "!" operator (like in line 840),
and
what happen if thres<0? (i'm new in spin language). The documentation in line_sensor() says: "return the analog value of the line sensor on the selected side"